recommendationAlgo/requirements.txt
sentence-transformers==2.2.2
torch>=1.9.0
supabase==2.0.0
python-dotenv==1.0.0
numpy>=1.21.0
scikit-learn>=1.0.0 